LayoutTests/webaudio/mediastreamaudiosourcenode-expected.txt

Basic tests for MediaStreamAudioSourceNode API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS {audio:true} generated stream
PASS stream.getAudioTracks().length is 1
PASS stream.getVideoTracks().length is 0
PASS Source AudioNode has no inputs.
PASS Source AudioNode has one output.
PASS connect() exception {"TypeError", "Argument 1 ('destination') to AudioNode.connect must be an instance of AudioNode"} thrown for illegal destination AudioNode
PASS connect() exception {"IndexSizeError", "The index is not in the allowed range."} thrown for illegal output index
PASS connect() exception {"IndexSizeError", "The index is not in the allowed range."} thrown for illegal input index
PASS mediaStreamSource.connect(context.destination) succeeded.
PASS onaudioprocess was called
PASS successfullyParsed is true

TEST COMPLETE
